Refactors question validators tests passing.
